move plugin to core, move arrp debug out of support helper
